update tabularize test; save tabularize column as tuple
